requirements.txt
email_validator==2.2.0
fastapi==0.115.0
fastapi-cli==0.0.5
httptools==0.6.1
httpx==0.27.2
itsdangerous==2.2.0
Jinja2==3.1.4
orjson==3.10.7
pip==24.2
pydantic-extra-types==2.9.0
pydantic-settings==2.5.2
python-multipart==0.0.10
PyYAML==6.0.2
SQLAlchemy==2.0.35
ujson==5.10.0
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
